# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.CategoryRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - name
    type: object
...
  ecommerce-backend_internal_delivery_http.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/http.Error'
      message:
        type: string
      meta:
        $ref: '#/definitions/http.Meta'
      success:
        type: boolean
    type: object
  http.Error:
    properties:
      code:
        type: string
      details: {}
    type: object
  http.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the backend API for E-Commerce platform
  title: E-Commerce API
  version: "1.0"
paths:
...
  /categories:
    get:
      description: Get a list of all product categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
      summary: List all categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create a new product category. Requires admin role.
      parameters:
      - description: Category Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - categories
  /categories/{id}:
    get:
      description: Get a product category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
      summary: Get category details
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Update an existing product category. Requires admin role.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ecommerce-backend_internal_delivery_http.Response'
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - categories
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"